fix(insights): solved connectivity between microservices and model integrity

+++ Insights.Gateway/Model/GeoInfoDto.cs
@@ -11,12 +11,12 @@
         public string Subregion { get; set; } = string.Empty;
         public string Flag { get; set; } = string.Empty;
         public string WeatherDescription { get; set; } = string.Empty;
-        public float Temperature { get; set; }
-        public float MaxTemperature { get; set; }
-        public float MinTemperature { get; set; }
-        public float FeelLike { get; set; }
-        public float Humidity { get; set; }
-        public float WindSpeed { get; set; }
+        public double Temperature { get; set; }
+        public double MaxTemperature { get; set; }
+        public double MinTemperature { get; set; }
+        public double FeelLike { get; set; }
+        public double Humidity { get; set; }
+        public double WindSpeed { get; set; }
         public DateTime Sunrise { get; set; }
         public DateTime Sunset { get; set; }
     }
